requirements.txt
aiohttp==3.9.1
aiohttp-client-cache==0.10.0
aiosignal==1.3.1
annotated-types==0.6.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
braceexpand==0.1.7
click==8.1.7
disnake==2.9.2
disnake-jishaku==2.6.5
frozenlist==1.4.1
idna==3.6
import_expression==1.1.5
itsdangerous==2.1.2
multidict==6.0.4
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.1
six==1.16.0
typing_extensions==4.9.0
url-normalize==1.4.3
wattpad==1.0.0
yarl==1.9.4
